# api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: microservice Api
  version: 0.1.9
  description: This is a microservice api.
servers: 
 - url: http://localhost:3000/v1
paths:
...
  /image_thumbnail:
    post:
      tags:
        - thumbnail
      summary: Thumbnail Generator
      description: this endpoint accepts a public url and returns the resulting thumbnail
      operationId: thumbnail
      security:
        - bearerAuth: []
      requestBody:
        description: accepts image url parameter*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Thumbnail'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'success'
                  thumbnail:
                    type: string
                    format: binary
                    example: hdfjhjdkjsjjgu-dfgnjdnjgkfjgd-sdnfgjdfnjdk
        "401":
          description: token Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: no image url found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: Invalid image url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Thumbnail:
      type: object
      properties:
        imageUrl:
          type: string
          example: https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__340.jpg
...
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        message: 
          type: string
          example: error message
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT